editing after code review
