+++ src/test/java/site/nomoreparties/stellarburgers/RedirectFromLkIntoMainPageTest.java
@@ -9,6 +9,7 @@ import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.junit.runners.Parameterized;
 import org.openqa.selenium.By;
+import site.nomoreparties.stellarburgers.steps.*;
 
 import static site.nomoreparties.stellarburgers.ChoosingBrowser.*;
 import static site.nomoreparties.stellarburgers.pom.constants.locators.AccountPageLocators.CONSTRUCTOR_BUTTON;
@@ -17,9 +18,14 @@ import static site.nomoreparties.stellarburgers.pom.constants.locators.MainPageL
 
 @DisplayName("Переход из личного кабинета в конструктор")
 @RunWith(Parameterized.class)
-public class RedirectFromLkIntoMainPageTest extends Steps{
+public class RedirectFromLkIntoMainPageTest {
+    private GeneralSteps generalSteps = new GeneralSteps();
+    private MainPageSteps mainPageSteps = new MainPageSteps();
+    private AccountPageSteps accountPageSteps = new AccountPageSteps();
+    private LoginPageSteps loginPageSteps = new LoginPageSteps();
     private final By mainPageButton;
     private final String buttonName;
+
 
     public RedirectFromLkIntoMainPageTest(By mainPageButton, String buttonName) {
         this.mainPageButton = mainPageButton;
@@ -36,24 +42,26 @@ public class RedirectFromLkIntoMainPageTest extends Steps{
 
     @Before
     public void setUp() {
-        setUpDriver();
-        accessToken = createUser();
+        generalSteps.setUpDriver();
+        generalSteps.accessToken = generalSteps.createUser();
     }
     @Test
     @DisplayName("Параметризованный тест на переход из личного кабинета на главную страницу с конструктором")
     @Description("Проверка что авторизованный пользователь может перейти из личного кабинета на главную страницу с конструктором")
     public void redirectFromLkIntoMainPage() {
-        clickOnEnterButtonOnMainPage();
-        setDataOnLoginPage();
-        clickOnEnterButtonOnLoginPage();
-        clickOnLkButtonOnMainPage();
-        waitForAccountPage();
-        clickOnMainPageButtonOnAccountPage(mainPageButton);
+        mainPageSteps.clickOnEnterButtonOnMainPage();
+        loginPageSteps.setDataOnLoginPage();
+        loginPageSteps.clickOnEnterButtonOnLoginPage();
+        mainPageSteps.waitForMainPage();
+        mainPageSteps.clickOnLkButtonOnMainPage();
+        accountPageSteps.waitForAccountPage();
+        accountPageSteps.clickOnMainPageButtonOnAccountPage(mainPageButton);
+        mainPageSteps.waitForMainPage();
         Assert.assertTrue(driver.findElement(MAKE_ORDER_BUTTON).isEnabled());
     }
     @After
     public void shutDownAndDeleteUser() {
-        shutDownDriver();
-        deleteUser();
+        generalSteps.shutDownDriver();
+        generalSteps.deleteUser();
     }
 }
