+++ src/test/java/site/nomoreparties/stellarburgers/RedirectToLkTest.java
@@ -6,32 +6,44 @@ import org.hamcrest.MatcherAssert;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
+import site.nomoreparties.stellarburgers.pom.AccountPage;
+import site.nomoreparties.stellarburgers.steps.*;
 
 import static org.hamcrest.CoreMatchers.is;
+import static site.nomoreparties.stellarburgers.ChoosingBrowser.driver;
 import static site.nomoreparties.stellarburgers.pom.constants.testdata.UserTestData.*;
 
 @DisplayName("Переход в личный кабинет")
-public class RedirectToLkTest extends Steps{
+public class RedirectToLkTest {
+    private GeneralSteps generalSteps = new GeneralSteps();
+    private MainPageSteps mainPageSteps = new MainPageSteps();
+    private LoginPageSteps loginPageSteps = new LoginPageSteps();
+    private AccountPageSteps accountPageSteps = new AccountPageSteps();
+    private AccountPage accountPage;
+
+
     @Before
     public void setUp() {
-        setUpDriver();
-        accessToken = createUser();
+        generalSteps.setUpDriver();
+        generalSteps.accessToken = generalSteps.createUser();
+        accountPage = new AccountPage(driver);
     }
 
     @Test
     @DisplayName("Тест на переход в личный кабинет")
     @Description("Проверка что авторизованный пользователь может перейти в личный кабинет")
     public void redirectToLk() {
-        clickOnEnterButtonOnMainPage();
-        setDataOnLoginPage();
-        clickOnEnterButtonOnLoginPage();
-        clickOnLkButtonOnMainPage();
-        waitForAccountPage();
+        mainPageSteps.clickOnEnterButtonOnMainPage();
+        loginPageSteps.setDataOnLoginPage();
+        loginPageSteps.clickOnEnterButtonOnLoginPage();
+        mainPageSteps.waitForMainPage();
+        mainPageSteps.clickOnLkButtonOnMainPage();
+        accountPageSteps.waitForAccountPage();
         MatcherAssert.assertThat(accountPage.getActualEmailText(), is(EMAIL));
     }
     @After
     public void shutDownAndDeleteUser() {
-        shutDownDriver();
-        deleteUser();
+        generalSteps.shutDownDriver();
+        generalSteps.deleteUser();
     }
 }
