+++ src/test/java/site/nomoreparties/stellarburgers/RegistrationTest.java
@@ -6,46 +6,55 @@ import org.junit.After;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
+import site.nomoreparties.stellarburgers.pom.RegisterPage;
+import site.nomoreparties.stellarburgers.steps.*;
 
 import static site.nomoreparties.stellarburgers.ChoosingBrowser.*;
 import static site.nomoreparties.stellarburgers.pom.constants.locators.LoginPageLocators.ENTER_BUTTON;
 import static site.nomoreparties.stellarburgers.pom.constants.testdata.UserTestData.*;
 
 @DisplayName("Регистрация пользователя")
-public class RegistrationTest extends Steps{
+public class RegistrationTest {
 
     private static final String EXPECTED_ERROR_TEXT = "Некорректный пароль";
+    private GeneralSteps generalSteps = new GeneralSteps();
+    private MainPageSteps mainPageSteps = new MainPageSteps();
+    private LoginPageSteps loginPageSteps = new LoginPageSteps();
+    private RegisterPageSteps registerPageSteps = new RegisterPageSteps();
+    private RegisterPage registerPage;
 
     @Before
     public void setUp() {
-        setUpDriver();
+        generalSteps.setUpDriver();
+        registerPage = new RegisterPage(driver);
     }
 
     @Test
     @DisplayName("Успешная регистрация пользователя")
     @Description("Проверка что пользователь может зарегистрироваться")
     public void registerAccountSuccessfully() {
-        clickOnEnterButtonOnMainPage();
-        clickOnRegisterButtonOnLoginPage();
-        setDataOnRegisterPage(VALID_PASSWORD);
-        clickOnRegisterButtonOnRegisterPage();
+        mainPageSteps.clickOnEnterButtonOnMainPage();
+        loginPageSteps.clickOnRegisterButtonOnLoginPage();
+        registerPageSteps.setDataOnRegisterPage(VALID_PASSWORD);
+        registerPageSteps.clickOnRegisterButtonOnRegisterPage();
+        loginPageSteps.waitForLoginPage();
         Assert.assertTrue(driver.findElement(ENTER_BUTTON).isEnabled());
     }
     @Test
     @DisplayName("Попытка регистрации пользователя с некорректным паролем")
     @Description("Проверка что пользователь не может зарегистрироваться и получение ошибки")
     public void registerAccountInvalidPassword() {
-        clickOnEnterButtonOnMainPage();
-        clickOnRegisterButtonOnLoginPage();
-        setDataOnRegisterPage(INVALID_PASSWORD);
-        clickOnNameFieldOnRegisterPage();
+        mainPageSteps.clickOnEnterButtonOnMainPage();
+        loginPageSteps.clickOnRegisterButtonOnLoginPage();
+        registerPageSteps.setDataOnRegisterPage(INVALID_PASSWORD);
+        registerPageSteps.clickOnNameFieldOnRegisterPage();
         Assert.assertEquals(EXPECTED_ERROR_TEXT, registerPage.getActualPasswordErrorText());
     }
 
     @After
     public void shutDownAndDeleteUser() {
-        shutDownDriver();
-        accessToken = deleteUserPreparation();
-        deleteUser();
+        generalSteps.shutDownDriver();
+        generalSteps.accessToken = generalSteps.deleteUserPreparation();
+        generalSteps.deleteUser();
     }
 }
